requirements.txt
pyttsx3
python-dotenv
openai
speechrecognition
PyAudio
unstructured